pyproject.toml
[project]
name = "ditto"
version = "1.1.11"
description = "Server to render quotes from a Notion database to images."
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.127.0",
    "loguru>=0.7.3",
    "notion-client==2.5.0",
    "numpy>=2.4.0",
    "pillow>=12.0.0",
    "pydantic-settings>=2.13.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
    "wand>=0.6.13",
]

[dependency-groups]
dev = ["pytest"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
# Add basic rules: E (pycodestyle errors) and F (Pyflakes).
select = ["E", "F"]
ignore = []

[tool.ruff.format]
# specific formatting options if needed, otherwise it inherits line-length
quote-style = "double"
indent-style = "space"
docstring-code-format = true
docstring-code-line-length = "dynamic"
